+++ skysail.server.app.plugins/src/io/skysail/server/app/plugins/obr/ObrRepository.java
@@ -47,8 +47,4 @@ public class ObrRepository implements Identifiable {
                 .collect(Collectors.toList());
     }
 
-    @Override
-    public void setId(String id) {
-    }
-
 }