+++ skysail.server.app.plugins/src/io/skysail/server/app/plugins/obr/ObrResource.java
@@ -1,10 +1,10 @@
 package io.skysail.server.app.plugins.obr;
 
+import org.apache.felix.bundlerepository.Resource;
+
 import io.skysail.domain.Identifiable;
 import io.skysail.domain.html.Field;
 import lombok.Getter;
-
-import org.apache.felix.bundlerepository.Resource;
 
 @Getter
 public class ObrResource implements Identifiable {
@@ -48,10 +48,6 @@ public class ObrResource implements Identifiable {
 
     public void setSearchFor(String searchFor) {
         this.searchFor = searchFor;
-    }
-
-    @Override
-    public void setId(String id) {
     }
 
 }