+++ skysail.server.converter/test/io/skysail/server/converter/impl/test/AnEntity.java
@@ -9,8 +9,4 @@ public class AnEntity implements Identifiable {
         return null;
     }
 
-    @Override
-    public void setId(String id) {
-    }
-
 }