+++ skysail.server.text.store.bundleresource/src/io/skysail/server/text/store/bundleresource/impl/Message.java
@@ -90,8 +90,4 @@ public class Message implements Identifiable {
         return msgKey;
     }
 
-    @Override
-    public void setId(String id) {
-        // do nothing
-    }
 }