+++ skysail.server/src/io/skysail/server/menus/MenuItemDescriptor.java
@@ -27,8 +27,5 @@ public class MenuItemDescriptor implements Identifiable {
         return url;
     }
 
-    @Override
-    public void setId(String id) {
-    }
 
 }