+++ skysail.server/src/io/skysail/server/menus/package-info.java
@@ -1,4 +1,4 @@
-@Version("1.2.0")
+@Version("2.0.0")
 package io.skysail.server.menus;
 
 import org.osgi.annotation.versioning.Version;