+++ skysail.server/test/io/skysail/server/utils/test/TestListServerResource.java
@@ -1,12 +1,13 @@
 package io.skysail.server.utils.test;
 
+import java.util.Arrays;
+import java.util.List;
+
+import org.mockito.Mockito;
+
 import io.skysail.domain.Identifiable;
 import io.skysail.server.app.SkysailApplication;
 import io.skysail.server.restlet.resources.ListServerResource;
-
-import java.util.*;
-
-import org.mockito.Mockito;
 
 public class TestListServerResource extends ListServerResource<Identifiable> {
 
@@ -17,10 +18,6 @@ public class TestListServerResource extends ListServerResource<Identifiable> {
             @Override
             public String getId() {
                 return null;
-            }
-
-            @Override
-            public void setId(String id) {
             }
 
         });